requirements.txt
anyio @ file:///C:/ci/anyio_1644463701441/work/dist
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///C:/ci/argon2-cffi-bindings_1644569878360/work
astroid==2.5.8
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
attrs==22.2.0
Automat==22.10.0
autopep8==1.5.6
Babel @ file:///C:/b/abs_a2shv_3tqi/croot/babel_1671782804377/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///C:/ci/beautifulsoup4_1650274792587/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
brotlipy==0.7.0
certifi @ file:///C:/b/abs_55jxithrm3/croot/certifi_1717618077850/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
comm @ file:///C:/b/abs_1419earm7u/croot/comm_1671231131638/work
constantly==15.1.0
contourpy==1.0.7
cryptography==38.0.1
cssselect==1.2.0
cycler==0.11.0
debugpy @ file:///C:/ci/debugpy_1637073815078/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
dnspython==1.16.0
docker==6.0.1
entrypoints @ file:///C:/ci/entrypoints_1649926621247/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastjsonschema @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_ebruxzvd08/croots/recipe/python-fastjsonschema_1661376484940/work
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
greenlet==3.0.3
hyperlink==21.0.0
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1648544472910/work
importlib-resources @ file:///tmp/build/80754af9/importlib_resources_1625135880749/work
incremental==22.10.0
ipykernel @ file:///C:/b/abs_b4f07tbsyd/croot/ipykernel_1672767104060/work
ipython @ file:///C:/b/abs_254uk73z5b/croot/ipython_1691532131313/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
isort==5.11.4
itemadapter==0.7.0
itemloaders==1.0.6
jedi==0.18.2
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
jmespath==1.0.1
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonschema @ file:///C:/b/abs_59eyhnbyej/croots/recipe/jsonschema_1663375476535/work
jupyter-server @ file:///C:/b/abs_1cfi3__jl8/croot/jupyter_server_1671707636383/work
jupyter_client @ file:///C:/b/abs_b9pns5mx5p/croot/jupyter_client_1671703062216/work
jupyter_core @ file:///C:/b/abs_84df679bho/croot/jupyter_core_1672332237650/work
jupyterlab @ file:///C:/b/abs_36s1nw_wru/croot/jupyterlab_1672132698305/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab_server @ file:///C:/b/abs_5ayr1vni0w/croot/jupyterlab_server_1672127362727/work
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
lxml==4.9.2
MarkupSafe @ file:///C:/ci/markupsafe_1654489871526/work
matplotlib==3.6.3
matplotlib-inline @ file:///C:/ci/matplotlib-inline_1661934035815/work
mccabe==0.6.1
mistune==0.8.4
mysql-connector-python==8.4.0
nbclassic @ file:///C:/b/abs_26e3fkk516/croot/nbclassic_1668174974037/work
nbclient @ file:///C:/ci/nbclient_1650290386732/work
nbconvert @ file:///C:/b/abs_4av3q4okro/croot/nbconvert_1668450658054/work
nbformat @ file:///C:/b/abs_85_3g7dkt4/croot/nbformat_1670352343720/work
nest-asyncio @ file:///C:/b/abs_3a_4jsjlqu/croot/nest-asyncio_1672387322800/work
notebook @ file:///C:/b/abs_ca13hqvuzw/croot/notebook_1668179888546/work
notebook_shim @ file:///C:/b/abs_ebfczttg6x/croot/notebook-shim_1668160590914/work
numpy==1.24.1
oauthlib==3.2.2
packaging==23.0
pandas==1.5.3
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parsel==1.7.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.2.0
pkgutil_resolve_name @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_81wm45v3kb/croots/recipe/pkgutil-resolve-name_1661463352381/work
platformdirs @ file:///C:/b/abs_73cc5cz_1u/croots/recipe/platformdirs_1662711386458/work
playwright==1.44.0
prometheus-client @ file:///C:/Windows/TEMP/abs_ab9nx8qb08/croots/recipe/prometheus_client_1659455104602/work
prompt-toolkit @ file:///C:/b/abs_6coz5_9f2s/croot/prompt-toolkit_1672387908312/work
Protego==0.2.1
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyDispatcher==2.0.6
pyee==11.1.0
Pygments==2.14.0
pylint==2.7.2
pymongo==3.11.3
PyMySQL==1.1.1
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent @ file:///C:/ci/pyrsistent_1636111468851/work
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==0.21.1
pytz @ file:///C:/b/abs_22fofvpn1x/croot/pytz_1671698059864/work
pywin32==305.1
pywinpty @ file:///C:/ci_310/pywinpty_1644230983541/work/target/wheels/pywinpty-2.0.2-cp38-none-win_amd64.whl
PyYAML==6.0
pyzmq @ file:///C:/ci/pyzmq_1657616005830/work
queuelib==1.6.2
requests==2.28.2
requests-oauthlib==1.3.1
retrying==1.3.4
scrapinghub==2.4.0
Scrapy==2.4.1
scrapy-playwright==0.0.36
scrapy-selenium==0.0.7
selenium==3.141.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
service-identity==21.1.0
shub==2.14.3
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///C:/ci/sniffio_1614030707456/work
soupsieve @ file:///C:/b/abs_fasraqxhlv/croot/soupsieve_1666296394662/work
SQLAlchemy==2.0.31
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
terminado @ file:///C:/b/abs_25nakickad/croot/terminado_1671751845491/work
tinycss2 @ file:///C:/b/abs_52w5vfuaax/croot/tinycss2_1668168823131/work
toml==0.10.2
tomli @ file:///C:/Windows/TEMP/abs_ac109f85-a7b3-4b4d-bcfd-52622eceddf0hy332ojo/croots/recipe/tomli_1657175513137/work
tornado @ file:///C:/ci/tornado_1662476991259/work
tqdm==4.55.1
traitlets==5.8.1
tweepy==4.14.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing_extensions==4.12.2
urllib3==1.26.15
w3lib==2.1.1
wcwidth==0.2.6
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==1.4.2
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
wrapt==1.12.1
zipp @ file:///C:/b/abs_b9jfdr908q/croot/zipp_1672387552360/work
zope.interface==5.5.2
